requirements.txt
# Core dependencies
Flask==2.2.5
Werkzeug==2.2.3
pika==1.3.1
playwright==1.35.0
requests==2.31.0
python-dotenv==1.0.0
pymongo==4.3.3

# Async support
aiohttp==3.8.5
asyncio==3.4.3

# Retry and backoff
backoff==2.2.1

# Task scheduling
apscheduler==3.10.1

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1

# Firecrawl
firecrawl-py==1.11.0
